refactor api data boundaries
Refs #197
