+++ docs/architecture.md
@@ -40,6 +40,18 @@ The app uses same-origin Vercel paths for upstream aviation sources that are not
 
 These paths are implemented as Next.js Route Handlers under `src/app/api/proxy/**`. The handlers keep upstream access same-origin, validate route and query parameters, apply lightweight per-IP rate limits, reject disallowed browser origins, and cap upstream response body sizes before parsing.
 
+### Server boundary convention
+
+API routes under `src/app/api/**/route.js` are HTTP adapters. They parse request parameters, enforce proxy/security policy, call a domain mechanism, and translate the mechanism result into `Response` or `NextResponse`.
+
+Server-side business flow lives under `src/server/<domain>/`:
+
+- `<domain>.mechanism.js` owns source selection, fallback order, cache policy, request parameterization, and provider orchestration.
+- `<domain>.models.js` owns domain constants, result metadata, and mechanism-specific error types.
+- `<domain>.utils.js` owns pure normalization and predicate helpers.
+
+Persistence boundaries live under `src/app/api/dao/*.dao.js`. DAO files should contain Supabase/SQL reads and writes only; they should not choose providers, cache policy, fallback behavior, or import mechanism files. Compatibility barrels may remain under `src/services/**` for stable client/service imports, but new API route work should depend on the DAO/mechanism split.
+
 The nearby-airport proxy can also use Supabase as a persistent response cache. When `NEXT_PUBLIC_SUPABASE_URL` and `NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY` are configured in Vercel, `/api/proxy/airports/nearby` reads and writes `public.nearby_airport_cache` records with a 90-day `expires_at` TTL. The migration grants the `anon` role only the select/insert/update permissions needed for this public cache table, with RLS policies restricted to `nearby-airports:%` cache keys. Cache failures are non-fatal: the handler falls back to AIRAC and logs the Supabase read/write error server-side.
 
 Airport search and detail go directly to the OurAirports tables (`public.airports`, `public.runways`, `public.airport_frequencies`, `public.navaids`). The legacy `public.airport_metadata_cache` table from the old airportsapi.com client has been dropped — no live reader remained after the OurAirports cutover.
