+++ src/app/api/airport/[ident]/route.js
@@ -5,17 +5,24 @@ import {
   enforceProxyRequest,
   jsonProxyResponse,
 } from "@/services/apiProxySecurity.js";
-import { createAirportPageDataServiceFromEnv } from "@/services/airports/airportPageDataService.js";
-import { scheduleRefreshIfDue } from "@/services/ourairports/ourAirportsRefresh.js";
+import {
+  getAirportDirectoryPage,
+  refreshAirportDirectoryIfDue,
+} from "@/server/airport-directory/airportDirectory.mechanism.js";
+import {
+  AIRPORT_DIRECTORY_CACHE_HEADERS,
+  AirportDirectoryConfigurationError,
+} from "@/server/airport-directory/airportDirectory.models.js";
+import {
+  isValidAirportIdent,
+  normalizeAirportDetailOptions,
+  normalizeAirportIdent,
+} from "@/server/airport-directory/airportDirectory.utils.js";
 
 const rateLimit = {
   key: "api:airport-detail",
   maxRequests: 60,
   windowMs: 60_000,
-};
-
-const cacheHeaders = {
-  "Cache-Control": "public, s-maxage=300, stale-while-revalidate=3600",
 };
 
 export const runtime = "nodejs";
@@ -29,8 +36,8 @@ export async function GET(request, { params }) {
   if (securityResponse) return securityResponse;
 
   const resolvedParams = await params;
-  const ident = String(resolvedParams?.ident || "").trim().toUpperCase();
-  if (!/^[A-Z0-9]{2,7}$/.test(ident)) {
+  const ident = normalizeAirportIdent(resolvedParams?.ident);
+  if (!isValidAirportIdent(ident)) {
     return jsonProxyResponse(
       request,
       { error: "Invalid airport ident" },
@@ -39,46 +46,32 @@ export async function GET(request, { params }) {
   }
 
   const url = new URL(request.url);
-  const parseOptionalInt = (key) => {
-    const raw = url.searchParams.get(key);
-    if (raw == null || raw === "") return undefined;
-    const value = Number(raw);
-    return Number.isFinite(value) ? value : undefined;
-  };
-  const radiusRaw = parseOptionalInt("nearbyRadiusNm");
-  const radiusNm =
-    radiusRaw === undefined ? undefined : Math.max(1, Math.min(radiusRaw, 250));
-  const limitRaw = parseOptionalInt("nearbyLimit");
-  const nearbyLimit =
-    limitRaw === undefined ? undefined : Math.max(1, Math.min(limitRaw, 50));
-
-  const service = createAirportPageDataServiceFromEnv();
-  if (!service) {
-    return jsonProxyResponse(
-      request,
-      { error: "Airport database is not configured" },
-      { status: 503 },
-    );
-  }
+  const { radiusNm, nearbyLimit } = normalizeAirportDetailOptions({
+    nearbyRadiusNm: url.searchParams.get("nearbyRadiusNm"),
+    nearbyLimit: url.searchParams.get("nearbyLimit"),
+  });
 
   try {
-    const data = await service.getAirportPageData(ident, { radiusNm, nearbyLimit });
-    // Trigger a background OurAirports refresh if the cached data is past
-    // its TTL. Runs after the response is sent so the user never waits.
-    after(() => scheduleRefreshIfDue());
-    if (!data.airport) {
+    const data = await getAirportDirectoryPage({ ident, radiusNm, nearbyLimit });
+    after(() => refreshAirportDirectoryIfDue());
+    if (!data) {
       return jsonProxyResponse(
         request,
         { error: "Airport not found" },
         { status: 404 },
       );
     }
-    return jsonProxyResponse(
-      request,
-      { ...data, source: "ourairports" },
-      { headers: cacheHeaders },
-    );
+    return jsonProxyResponse(request, data, {
+      headers: AIRPORT_DIRECTORY_CACHE_HEADERS,
+    });
   } catch (error) {
+    if (error instanceof AirportDirectoryConfigurationError) {
+      return jsonProxyResponse(
+        request,
+        { error: error.message },
+        { status: error.status },
+      );
+    }
     console.error("[api/airport] failed", error);
     return jsonProxyResponse(
       request,
