+++ src/app/api/proxy/aircraft/photos/[hex]/route.js
@@ -4,8 +4,18 @@ import {
   enforceProxyRequest,
   jsonProxyResponse,
   normalizeAircraftHex,
-  readResponseJson,
 } from "@/services/apiProxySecurity.js";
+import {
+  getAircraftPhoto,
+} from "@/server/aircraft-photos/aircraftPhotos.mechanism.js";
+import {
+  AIRCRAFT_PHOTO_CACHE_HEADERS,
+  AIRCRAFT_PHOTO_SOURCE,
+  AircraftPhotoProviderError,
+} from "@/server/aircraft-photos/aircraftPhotos.models.js";
+import {
+  buildAircraftPhotoQuery,
+} from "@/server/aircraft-photos/aircraftPhotos.utils.js";
 
 const rateLimit = {
   key: "proxy:aircraft-photos",
@@ -13,98 +23,8 @@ const rateLimit = {
   windowMs: 60_000,
 };
 
-const USER_AGENT = "ADSBao/0.11.0 (+https://github.com/orriduck/ADSBao)";
-const PHOTO_MAX_BYTES = 256 * 1024;
-const PHOTO_SOURCE = "planespotters.net";
-
 export function OPTIONS(request) {
   return createCorsPreflightResponse(request);
-}
-
-function sanitizeOptionalCode(value, { max = 16 } = {}) {
-  const normalized = String(value || "").trim().toUpperCase();
-  return /^[A-Z0-9-]+$/.test(normalized) && normalized.length <= max
-    ? normalized
-    : "";
-}
-
-function buildPhotoUrl({ hex, request }) {
-  const requestUrl = new URL(request.url);
-  const url = new URL(
-    `https://api.planespotters.net/pub/photos/hex/${encodeURIComponent(hex)}`,
-  );
-  const registration = sanitizeOptionalCode(
-    requestUrl.searchParams.get("registration"),
-    { max: 12 },
-  );
-  const type = sanitizeOptionalCode(requestUrl.searchParams.get("type"), {
-    max: 8,
-  });
-
-  if (registration) url.searchParams.set("reg", registration);
-  if (type) url.searchParams.set("icaoType", type);
-  return url;
-}
-
-function buildImageProxyUrl({ hex, request }) {
-  const requestUrl = new URL(request.url);
-  const url = new URL(
-    `/api/proxy/aircraft/photos/${encodeURIComponent(hex)}/image`,
-    requestUrl.origin,
-  );
-  for (const key of ["registration", "type"]) {
-    const value = requestUrl.searchParams.get(key);
-    if (value) url.searchParams.set(key, value);
-  }
-  return `${url.pathname}${url.search}`;
-}
-
-function selectPhoto(payload) {
-  const [photo] = Array.isArray(payload?.photos) ? payload.photos : [];
-  const image = photo?.thumbnail_large || photo?.thumbnail;
-  const src = typeof image?.src === "string" ? image.src : "";
-  if (!src) return null;
-
-  return {
-    src: buildImageProxyUrl({ hex: payload?.hex, request: payload?.request }),
-    originalSrc: src,
-    width: Number(image?.size?.width) || null,
-    height: Number(image?.size?.height) || null,
-    link: typeof photo?.link === "string" ? photo.link : "",
-    photographer:
-      typeof photo?.photographer === "string" ? photo.photographer : "",
-    source: PHOTO_SOURCE,
-  };
-}
-
-async function fetchAircraftPhoto({ hex, request }) {
-  const url = buildPhotoUrl({ hex, request });
-
-  let response;
-  try {
-    response = await fetch(url, {
-      headers: {
-        Accept: "application/json",
-        "User-Agent": USER_AGENT,
-      },
-      next: { revalidate: 3600 },
-    });
-  } catch (networkError) {
-    const error = new Error(`network: ${networkError.message}`);
-    throw error;
-  }
-
-  if (!response.ok) {
-    const error = new Error(`HTTP ${response.status}`);
-    error.status = response.status;
-    throw error;
-  }
-
-  const payload = await readResponseJson(response, {
-    label: `${PHOTO_SOURCE} aircraft photo response`,
-    maxBytes: PHOTO_MAX_BYTES,
-  });
-  return selectPhoto({ ...payload, hex, request });
 }
 
 export async function GET(request, { params }) {
@@ -122,7 +42,12 @@ export async function GET(request, { params }) {
   }
 
   try {
-    const photo = await fetchAircraftPhoto({ hex, request });
+    const requestUrl = new URL(request.url);
+    const photo = await getAircraftPhoto({
+      hex,
+      origin: requestUrl.origin,
+      ...buildAircraftPhotoQuery(requestUrl.searchParams),
+    });
     if (!photo) {
       return jsonProxyResponse(
         request,
@@ -130,8 +55,8 @@ export async function GET(request, { params }) {
         {
           status: 404,
           headers: {
-            "Cache-Control": "public, s-maxage=3600, stale-while-revalidate=86400",
-            "X-Data-Source": PHOTO_SOURCE,
+            ...AIRCRAFT_PHOTO_CACHE_HEADERS,
+            "X-Data-Source": AIRCRAFT_PHOTO_SOURCE,
           },
         },
       );
@@ -140,13 +65,11 @@ export async function GET(request, { params }) {
     return Response.json(
       { hex, photo },
       {
-        headers: buildProxyHeaders(request, {
-          "Cache-Control": "public, s-maxage=3600, stale-while-revalidate=86400",
-          "X-Data-Source": PHOTO_SOURCE,
-        }),
+        headers: buildProxyHeaders(request, AIRCRAFT_PHOTO_CACHE_HEADERS),
       },
     );
   } catch (error) {
+    if (!(error instanceof AircraftPhotoProviderError)) throw error;
     console.warn(
       "[aircraft-photo] planespotters.net failed",
       error.status ? `status=${error.status}` : error.message,
