+++ src/app/api/proxy/flight-routes/callsign/[callsign]/route.js
@@ -3,149 +3,21 @@ import {
   createCorsPreflightResponse,
   enforceProxyRequest,
   jsonProxyResponse,
-  readResponseJson,
 } from "@/services/apiProxySecurity.js";
-import { withAuditLogging } from "@/utils/apiLogger.js";
 import {
-  buildVrsRouteResponse,
-  buildVrsRouteUrl,
   normalizeRouteCallsign,
-  shouldUseAerodataboxFallback,
-  VRS_ROUTE_MISS_STATUS,
-  VRS_ROUTE_USER_AGENT,
 } from "@/services/aviation/vrsRouteProxyModel.js";
 import {
-  AERODATABOX_RAPIDAPI_HOST,
-  buildAerodataboxFlightRouteResponse,
-  buildAerodataboxFlightUrl,
-  reserveAerodataboxRequestSlot,
-  resolveAerodataboxDateLocal,
-  shouldSuppressVrsRouteAfterAerodataboxStatus,
-} from "@/services/aviation/aerodataboxRouteProxyModel.js";
-
-const aerodataboxRapidApiKey = process.env.AERODATABOX_RAPIDAPI_KEY || "";
-const aerodataboxRapidApiHost =
-  process.env.AERODATABOX_RAPIDAPI_HOST || AERODATABOX_RAPIDAPI_HOST;
-let nextAerodataboxRequestAt = 0;
-
-const ROUTE_CACHE_TTL_SECONDS = 60 * 60;
-const ROUTE_MISS_CACHE_TTL_SECONDS = 5 * 60;
-const ROUTE_STALE_WHILE_REVALIDATE_SECONDS = 10 * 60;
-
-export function buildRouteCacheHeaders(body) {
-  const ttl = body ? ROUTE_CACHE_TTL_SECONDS : ROUTE_MISS_CACHE_TTL_SECONDS;
-  const swr = ROUTE_STALE_WHILE_REVALIDATE_SECONDS;
-  const sharedValue = `public, s-maxage=${ttl}, stale-while-revalidate=${swr}`;
-  return {
-    "Cache-Control": `public, max-age=0, s-maxage=${ttl}, stale-while-revalidate=${swr}`,
-    "CDN-Cache-Control": sharedValue,
-    "Vercel-CDN-Cache-Control": sharedValue,
-  };
-}
-
-const sleep = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
-
-async function waitForAerodataboxSlot() {
-  const slot = reserveAerodataboxRequestSlot({
-    now: Date.now(),
-    nextAllowedAt: nextAerodataboxRequestAt,
-  });
-  nextAerodataboxRequestAt = slot.nextAllowedAt;
-  if (slot.delayMs > 0) await sleep(slot.delayMs);
-}
-
-async function fetchVrsStandingRoute(callsign) {
-  let response;
-  try {
-    response = await fetch(buildVrsRouteUrl(callsign), {
-      headers: {
-        "User-Agent": VRS_ROUTE_USER_AGENT,
-        Accept: "application/json",
-      },
-      signal: AbortSignal.timeout(9_000),
-    });
-  } catch (err) {
-    console.warn(`[vrs-route] fetch failed for ${callsign}:`, err.message);
-    return null;
-  }
-
-  if (response.status === 404) {
-    return null;
-  }
-
-  if (!response.ok) {
-    console.warn(`[vrs-route] HTTP ${response.status} for ${callsign}`);
-    return null;
-  }
-
-  const payload = await readResponseJson(response, {
-    label: "VRS standing-data route",
-    maxBytes: 512 * 1024,
-  });
-  return buildVrsRouteResponse(callsign, payload);
-}
-
-async function fetchAerodataboxRoute(callsign, targetAirport) {
-  if (!aerodataboxRapidApiKey) {
-    return { route: null, suppressVrsRoute: false };
-  }
-
-  const url = buildAerodataboxFlightUrl(callsign, resolveAerodataboxDateLocal());
-  const auditedFetch = withAuditLogging(
-    (requestUrl, options) => fetch(requestUrl, options),
-    { service: "aerodatabox/FlightStatus" },
-  );
-  let response;
-  try {
-    await waitForAerodataboxSlot();
-    response = await auditedFetch(url, {
-      headers: {
-        Accept: "application/json",
-        "X-RapidAPI-Key": aerodataboxRapidApiKey,
-        "X-RapidAPI-Host": aerodataboxRapidApiHost,
-      },
-      signal: AbortSignal.timeout(9_000),
-    });
-  } catch (err) {
-    console.warn(`[aerodatabox-route] fetch failed for ${callsign}:`, err.message);
-    return { route: null, suppressVrsRoute: false };
-  }
-
-  if (response.status === 204 || response.status === 404) {
-    return { route: null, suppressVrsRoute: false };
-  }
-
-  if (!response.ok) {
-    console.warn(`[aerodatabox-route] HTTP ${response.status} for ${callsign}`);
-    return {
-      route: null,
-      suppressVrsRoute: shouldSuppressVrsRouteAfterAerodataboxStatus(
-        response.status,
-      ),
-    };
-  }
-
-  const payload = await readResponseJson(response, {
-    label: "AeroDataBox flight status route",
-    maxBytes: 512 * 1024,
-  });
-  return {
-    route: buildAerodataboxFlightRouteResponse(
-      callsign,
-      payload,
-      targetAirport,
-    ),
-    suppressVrsRoute: false,
-  };
-}
-
-function getTargetAirport(request) {
-  const params = request.nextUrl?.searchParams;
-  return {
-    icao: params?.get("airportIcao") || params?.get("airport") || "",
-    iata: params?.get("airportIata") || "",
-  };
-}
+  buildRouteCacheHeaders,
+  VRS_ROUTE_MISS_STATUS,
+} from "@/server/flight-routes/flightRoutes.models.js";
+import {
+  resolveFlightRoute,
+} from "@/server/flight-routes/flightRoutes.mechanism.js";
+import {
+  getTargetAirportFromSearchParams,
+  shouldForceAerodatabox,
+} from "@/server/flight-routes/flightRoutes.utils.js";
 
 const rateLimit = {
   key: "proxy:flight-routes",
@@ -173,34 +45,14 @@ export async function GET(request, { params }) {
   }
 
   try {
-    const targetAirport = getTargetAirport(request);
-    const forceAerodatabox =
-      request.nextUrl?.searchParams.get("force") === "aerodatabox";
-
-    let body;
-    if (forceAerodatabox) {
-      // Explicit user-initiated revalidation: skip VRS standing-data and go
-      // straight to AeroDataBox so the freshest authoritative result lands
-      // even if our cache has the VRS answer already.
-      const aerodataboxResult = await fetchAerodataboxRoute(
-        callsign,
-        targetAirport,
-      );
-      body = aerodataboxResult.route;
-    } else {
-      body = await fetchVrsStandingRoute(callsign);
-      if (shouldUseAerodataboxFallback(body, targetAirport)) {
-        const aerodataboxResult = await fetchAerodataboxRoute(
-          callsign,
-          targetAirport,
-        );
-        if (aerodataboxResult.route) {
-          body = aerodataboxResult.route;
-        } else if (aerodataboxResult.suppressVrsRoute) {
-          body = null;
-        }
-      }
-    }
+    const forceAerodatabox = shouldForceAerodatabox(request.nextUrl?.searchParams);
+    const body = await resolveFlightRoute({
+      callsign,
+      targetAirport: getTargetAirportFromSearchParams(
+        request.nextUrl?.searchParams,
+      ),
+      forceAerodatabox,
+    });
 
     const cacheHeaders = forceAerodatabox
       ? { "Cache-Control": "no-store" }
