+++ src/app/api/proxy/local-weather/[lat]/[lon]/route.js
@@ -3,14 +3,17 @@ import {
   createCorsPreflightResponse,
   enforceProxyRequest,
   jsonProxyResponse,
-  readResponseJson,
 } from "@/services/apiProxySecurity.js";
 import {
-  buildOpenMeteoCurrentWeatherUrl,
-  isValidOpenMeteoCurrentPayload,
+  fetchLocalWeather,
+} from "@/server/local-weather/localWeather.mechanism.js";
+import {
+  LocalWeatherProviderError,
+} from "@/server/local-weather/localWeather.models.js";
+import {
   normalizeLatitudeParam,
   normalizeLongitudeParam,
-} from "@/services/aviation/localWeatherProxyModel.js";
+} from "@/server/local-weather/localWeather.utils.js";
 
 const rateLimit = {
   key: "proxy:local-weather",
@@ -38,45 +41,27 @@ export async function GET(request, { params }) {
     );
   }
 
-  const url = buildOpenMeteoCurrentWeatherUrl({ latitude, longitude });
-
   try {
-    const response = await fetch(url, {
-      headers: {
-        Accept: "application/json",
-        "User-Agent": "ADSBao/0.10.0 (https://github.com/orriduck/ADSBao)",
-      },
-      next: {
-        revalidate: 300,
-      },
+    const result = await fetchLocalWeather({ latitude, longitude });
+    return Response.json(result.payload, {
+      status: result.status,
+      headers: buildProxyHeaders(request),
     });
-
-    if (!response.ok) {
-      return jsonProxyResponse(
-        request,
-        { error: "Failed to load weather" },
-        { status: response.status },
-      );
-    }
-
-    const payload = await readResponseJson(response, {
-      label: "Open-Meteo current weather response",
-      maxBytes: 512 * 1024,
-    });
-
-    if (!isValidOpenMeteoCurrentPayload(payload)) {
+  } catch (error) {
+    if (error instanceof LocalWeatherProviderError && error.message === "Invalid weather payload") {
       return jsonProxyResponse(
         request,
         { error: "Invalid weather payload" },
         { status: 502 },
       );
     }
-
-    return Response.json(payload, {
-      status: response.status,
-      headers: buildProxyHeaders(request),
-    });
-  } catch (error) {
+    if (error instanceof LocalWeatherProviderError && error.status !== 502) {
+      return jsonProxyResponse(
+        request,
+        { error: "Failed to load weather" },
+        { status: error.status },
+      );
+    }
     console.error("[local-weather] load failed", error);
     return jsonProxyResponse(
       request,
