+++ src/app/api/proxy/metar/[icao]/route.js
@@ -4,16 +4,18 @@ import {
   enforceProxyRequest,
   jsonProxyResponse,
   normalizeIcao,
-  readResponseJson,
 } from "@/services/apiProxySecurity.js";
+import { fetchMetar } from "@/server/metar/metar.mechanism.js";
+import {
+  METAR_CACHE_HEADERS,
+  MetarProviderError,
+} from "@/server/metar/metar.models.js";
 
 const rateLimit = {
   key: "proxy:metar",
   maxRequests: 90,
   windowMs: 60_000,
 };
-
-const USER_AGENT = "ADSBao/0.10.0 (https://github.com/orriduck/ADSBao)";
 
 export function OPTIONS(request) {
   return createCorsPreflightResponse(request);
@@ -30,56 +32,26 @@ export async function GET(request, { params }) {
     return jsonProxyResponse(request, { error: "Invalid ICAO" }, { status: 400 });
   }
 
-  const url = new URL("https://aviationweather.gov/api/data/metar");
-  url.searchParams.set("ids", icao);
-  url.searchParams.set("format", "json");
-
   try {
-    const response = await fetch(url, {
-      headers: {
-        Accept: "application/json",
-        "User-Agent": USER_AGENT,
-      },
-      next: {
-        revalidate: 60,
-      },
+    const payload = await fetchMetar({ icao });
+    return Response.json(payload, {
+      headers: buildProxyHeaders(request, METAR_CACHE_HEADERS),
     });
-
-    if (!response.ok) {
-      return jsonProxyResponse(
-        request,
-        { error: "Failed to load METAR" },
-        { status: response.status },
-      );
-    }
-
-    if (response.status === 204) {
-      return Response.json([], {
-        headers: buildProxyHeaders(request, {
-          "Cache-Control": "public, s-maxage=60, stale-while-revalidate=300",
-        }),
-      });
-    }
-
-    const payload = await readResponseJson(response, {
-      label: "AviationWeather METAR response",
-      maxBytes: 512 * 1024,
-    });
-
-    if (!Array.isArray(payload)) {
+  } catch (error) {
+    if (error instanceof MetarProviderError && error.message === "Invalid METAR payload") {
       return jsonProxyResponse(
         request,
         { error: "Invalid METAR payload" },
         { status: 502 },
       );
     }
-
-    return Response.json(payload, {
-      headers: buildProxyHeaders(request, {
-        "Cache-Control": "public, s-maxage=60, stale-while-revalidate=300",
-      }),
-    });
-  } catch (error) {
+    if (error instanceof MetarProviderError && error.status !== 502) {
+      return jsonProxyResponse(
+        request,
+        { error: "Failed to load METAR" },
+        { status: error.status },
+      );
+    }
     console.error(`[metar] load failed for ${icao}`, error);
     return jsonProxyResponse(
       request,
