+++ src/server/aircraft-icons/aircraftIcons.models.js
@@ -1,0 +1,6 @@
+export const AIRCRAFT_ICON_FALLBACK_NAME = "arrow";
+export const AIRCRAFT_ICON_FALLBACK_SVG =
+  '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M12 2L16 20L12 17L8 20Z" fill="black"/></svg>';
+export const AIRCRAFT_ICON_MAX_BYTES = 64 * 1024;
+export const AIRCRAFT_ICON_CACHE_CONTROL =
+  "public, max-age=86400, s-maxage=86400, stale-while-revalidate=604800";
