+++ src/server/aircraft-icons/aircraftIcons.utils.js
@@ -1,0 +1,9 @@
+import { join } from "node:path";
+import process from "node:process";
+
+export const AIRCRAFT_ICON_DIR = join(
+  process.cwd(),
+  "public",
+  "icons",
+  "aircraft",
+);
