+++ src/server/aircraft-positions/aircraftPositions.utils.js
@@ -1,0 +1,4 @@
+export function formatAircraftPositionAttempt(providerId, error) {
+  if (!error) return `${providerId}:200`;
+  return `${providerId}:${error.status || "ERR"}`;
+}
