+++ src/server/aircraft-trace/aircraftTrace.utils.js
@@ -1,0 +1,2 @@
+export const formatAircraftTraceAttempt = (providerId, status) =>
+  `${providerId}:${status}`;
