+++ src/server/flight-routes/flightRoutes.utils.js
@@ -1,0 +1,9 @@
+export const getTargetAirportFromSearchParams = (searchParams) => ({
+  icao: searchParams?.get("airportIcao") || searchParams?.get("airport") || "",
+  iata: searchParams?.get("airportIata") || "",
+});
+
+export const shouldForceAerodatabox = (searchParams) =>
+  searchParams?.get("force") === "aerodatabox";
+
+export const sleep = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
