+++ src/server/local-weather/localWeather.utils.js
@@ -1,0 +1,4 @@
+export {
+  normalizeLatitudeParam,
+  normalizeLongitudeParam,
+} from "@/services/aviation/localWeatherProxyModel.js";
