+++ src/server/metar/metar.utils.js
@@ -1,0 +1,6 @@
+export function buildMetarUrl(icao) {
+  const url = new URL("https://aviationweather.gov/api/data/metar");
+  url.searchParams.set("ids", icao);
+  url.searchParams.set("format", "json");
+  return url;
+}
